Resolve 64 bit to 32 bit convertion
The code is now almost completely 64 bit safe for large files. The sole
exeption being two QByteArray contructor calls. The size limit for files
is 1GB until this is resolved.

+++ src/gnudiff_io.cpp
@@ -510,7 +510,7 @@ verify(enough_prime_offsets,
 
 bool GnuDiff::read_files(struct file_data filevec[], bool /*pretend_binary*/)
 {
-    int i;
+    qint64 i;
 
     find_identical_ends(filevec);
